+++ src/main/java/mod/kinderhead/luadatapack/datapack/ReloadListener.java
@@ -29,6 +29,8 @@ public class ReloadListener implements IdentifiableResourceReloadListener, Synch
         }
 
         LuaDatapack.LOGGER.info("Loaded " + String.valueOf(count) + " lua scripts");
+
+        Scripts.initAll();
     }
 
     @Override