+++ src/main/java/mod/kinderhead/luadatapack/datapack/Scripts.java
@@ -5,47 +5,54 @@ import java.util.HashMap;
 import java.util.Map;
 
 import mod.kinderhead.luadatapack.LuaDatapack;
+import net.minecraft.text.Text;
 import net.minecraft.util.Identifier;
 
 public class Scripts {
-    private static Map<Identifier, String> files = new HashMap<Identifier, String>();
+    private static Map<String, Project> projects = new HashMap<String, Project>();
 
     public static String get(Identifier id) {
-        for (var i : files.entrySet()) {
-            if (i.getKey().getNamespace().equals(id.getNamespace()) && i.getKey().getPath().equals(id.getPath())) {
-                return i.getValue();
-            }
+        if (projects.containsKey(id.getNamespace()) && projects.get(id.getNamespace()).has(id)) {
+            return projects.get(id.getNamespace()).get(id);
         }
 
         LuaDatapack.LOGGER.error("Could not find file with id " + id.toString());
         return null;
     }
 
+    public static String getImportable(Identifier id) {
+        if (projects.containsKey(id.getNamespace()) && projects.get(id.getNamespace()).getImportableIds().contains(id)) {
+            return projects.get(id.getNamespace()).get(id);
+        }
+
+        LuaDatapack.LOGGER.error("Could not find importable file with id " + id.toString() + ". Check project configuration if this is a mistake");
+        return null;
+    }
+
     public static void set(Identifier id, String code) {
-        files.put(id, code);
+        if (!projects.containsKey(id.getNamespace())) projects.put(id.getNamespace(), new Project(id.getNamespace()));
+        projects.get(id.getNamespace()).put(id, code);
     }
 
     public static void clear() {
-        files.clear();
+        projects.clear();
     }
 
     public static Iterable<Identifier> idSet() {
-        return files.keySet();
+        ArrayList<Identifier> ids = new ArrayList<>();
+
+        for (Project i : projects.values()) {
+            ids.addAll(i.getCallableIds());
+        }
+
+        return ids;
     }
 
-    public static Iterable<Identifier> idSet(boolean excludeStd) {
-        if (!excludeStd) {
-            return idSet();
+    public static void initAll() {
+        for (Project i : projects.values()) {
+            if (!i.init()) {
+                LuaDatapack.SERVER.getPlayerManager().broadcast(Text.translatable("text.luadatapack.project_error", i.name), false);
+            }
         }
-
-        ArrayList<Identifier> list = new ArrayList<>();
-
-        for (var i : files.keySet()) {
-            if (!i.getNamespace().equals("std")) {
-                list.add(i);
-            } 
-        }
-
-        return list;
     }
 }