+++ src/main/resources/data/std/lua/_default_project.lua
@@ -1,0 +1,6 @@
+local config = require("std:config")
+
+return {
+    callables = config.without_underscore(),
+    importables = {}
+}