+++ src/main/resources/data/std/lua/_project.lua
@@ -1,0 +1,6 @@
+local config = require("std:config")
+
+return {
+    callables = {},
+    importables = config.without_underscore()
+}